requirements.txt
kafka-python==2.2.11
requests==2.31.0
python-dotenv==1.0.0
pandas==2.3.0
pyarrow==20.0.0
fastparquet==2024.11.0
streamlit==1.46.0
streamlit-folium==0.25.0
folium==0.20.0
geopandas==1.0.1
pydeck==0.9.1
streamlit-autorefresh==1.0.1
SQLAlchemy==2.0.41
psycopg2-binary
GeoAlchemy2==0.17.1
alembic==1.16.4
